#14 Move shell to own file
